databuilder/requirements.txt
numpy
cython
elasticsearch>=6.2.0,<8.0
elasticsearch-dsl==7.4.0
neo4j-driver>=4.4.5,<5.0
requests>=2.25.0,<3.0

freezegun>=1.1.0
atomicwrites>=1.1.5
more-itertools>=4.2.0
pluggy>=0.6.0
py>=1.10.0
pyhocon>=0.3.42
pyparsing>=2.2.0
sqlalchemy>=1.3.6
wheel>=0.31.1
pytz>=2018.4
statsd>=3.2.1
retrying>=1.3.3
unicodecsv>=0.14.1,<1.0
httplib2>=0.18.0
unidecode
Jinja2>=2.10.0,<4
pandas>=0.21.0,<1.5.0
responses>=0.10.6
jsonref==0.2
protobuf==3.19.6
amundsen-common>=0.16.0
amundsen-rds==0.0.8
google-api-python-client
google-cloud-bigquery